// src/shared/types/generated/application/CreatedFolder.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FolderId } from "./FolderId";
import type { FolderMutationReceipt } from "./FolderMutationReceipt";

export type CreatedFolder = { folder_id: FolderId, receipt: FolderMutationReceipt, };